Users are now only able to delete items they put of for sale

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -1,17 +1,17 @@
 const bcrypt = require("bcryptjs");
-const jwt = require('jsonwebtoken');
-const { randomBytes } = require('crypto');
-const { promisify } = require('util');
-const { transport, makeANiceEmail } = require('../mail');
-const { hasPermission } = require('../utils');
+const jwt = require("jsonwebtoken");
+const { randomBytes } = require("crypto");
+const { promisify } = require("util");
+const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
     // check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error('You must be logged in to do that!');
+      throw new Error("You must be logged in to do that!");
     }
-    
+
     const item = await ctx.db.mutation.createItem(
       {
         data: {
@@ -48,9 +48,23 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
-    // 2. Check if they own that item , or have the permissions
-    // TODO
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+
+    // 2. Check if they are signed in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to do that!");
+    }
+
+    // 3. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You do not have permission to do that!');
+    } 
+
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -67,7 +81,7 @@ const Mutations = {
         data: {
           ...args,
           password,
-          permissions: { set: ['USER'] }
+          permissions: { set: ["USER"] }
         }
       },
       info
@@ -76,9 +90,9 @@ const Mutations = {
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
 
     // set the jwt as a cookie on the response
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true, // cannot access token via javascript
-      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
+      maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year cookie
     });
 
     // return the user to the browser
@@ -94,24 +108,24 @@ const Mutations = {
     // 2. Check if their password is correct
     const valid = await bcrypt.compare(password, user.password);
     if (!valid) {
-      throw new Error('Invalid Password');
+      throw new Error("Invalid Password");
     }
 
     // 3. generate the JWT Token
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
 
     // 4. Set the cookie with the token
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true, // cannot access token via javascript
-      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
+      maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year cookie
     });
 
     // 5. Return the user
     return user;
   },
   signout(parent, args, ctx, info) {
-    ctx.response.clearCookie('token');
-    return { message: 'Goodbye!' }; 
+    ctx.response.clearCookie("token");
+    return { message: "Goodbye!" };
   },
   async requestReset(parent, args, ctx, info) {
     // 1. Check if this is a real user
@@ -122,7 +136,7 @@ const Mutations = {
 
     // 2. Set a reset token and expiry on that user
     const randomBytesPromiseified = promisify(randomBytes);
-    const resetToken = (await randomBytesPromiseified(20)).toString('hex');
+    const resetToken = (await randomBytesPromiseified(20)).toString("hex");
     const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
     const res = await ctx.db.mutation.updateUser({
       where: { email: args.email },
@@ -131,32 +145,34 @@ const Mutations = {
 
     // 3. Email them the reset token
     const mailRes = await transport.sendMail({
-      from: 'brian@donotreply.com',
+      from: "brian@donotreply.com",
       to: user.email,
-      subject: 'Your Password Reset Token',
+      subject: "Your Password Reset Token",
       html: makeANiceEmail(`Your Password Reset Token is here! 
       \n\n 
-      <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
+      <a href="${
+        process.env.FRONTEND_URL
+      }/reset?resetToken=${resetToken}">Click Here to Reset</a>`)
     });
 
     // 4. Return the message
-    return { message: 'Thanks!'};
+    return { message: "Thanks!" };
   },
   async resetPassword(parent, args, ctx, info) {
     // 1. check if the password match
     if (args.password !== args.confirmPassword) {
-      throw new Error('Your passwords do not match!');
+      throw new Error("Your passwords do not match!");
     }
     // 2. check if its a legit reset token
     // 3. check if its expired
     const [user] = await ctx.db.query.users({
       where: {
         resetToken: args.resetToken,
-        resetTokenExpiry_gte: Date.now() - 3600000,
-      },
+        resetTokenExpiry_gte: Date.now() - 3600000
+      }
     });
     if (!user) {
-      throw new Error('This token is either invalid or expired!');
+      throw new Error("This token is either invalid or expired!");
     }
     // 4. hash their new password
     const password = await bcrypt.hash(args.password, 10);
@@ -168,41 +184,47 @@ const Mutations = {
         resetToken: null,
         resetTokenExpiry: null
       }
-    })
+    });
     // 6. Generate JWT
     const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
     // 7. Set the JWT cookie
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365
-    })
+    });
     // 8. return the new user
     return updatedUser;
   },
   async updatePermissions(parent, args, ctx, info) {
     // 1. Check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error('You must be logged in!');
+      throw new Error("You must be logged in!");
     }
     // 2. Query the current user
-    const currentUser = await ctx.db.query.user({ 
-      where: {
-        id: ctx.request.userId,
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
       },
-    }, info);
+      info
+    );
     // 3. Check if they have permissions to do this
-    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
     // 4. Update the permissions
-    return ctx.db.mutation.updateUser({
-      data: {
-        permissions: {
-          set: args.permissions,
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
         },
+        where: {
+          id: args.userId
+        }
       },
-      where: {
-        id: args.userId
-      }
-    }, info);
+      info
+    );
   }
 };
